+++ Drivers/STM32F4xx_HAL_Driver/Inc/stm32f4xx_hal_cortex.h
@@ -125,8 +125,8 @@ typedef struct
 /** @defgroup CORTEX_MPU_Region_Enable CORTEX MPU Region Enable
   * @{
   */
-#define  MPU_REGION_ENABLE     (0x01)
-#define  MPU_REGION_DISABLE    (0x00)
+#define  MPU_REGION_ENABLE     ((uint8_t)0x01)
+#define  MPU_REGION_DISABLE    ((uint8_t)0x00)
 /**
   * @}
   */
@@ -134,8 +134,8 @@ typedef struct
 /** @defgroup CORTEX_MPU_Instruction_Access CORTEX MPU Instruction Access
   * @{
   */
-#define  MPU_INSTRUCTION_ACCESS_ENABLE      (0x00)
-#define  MPU_INSTRUCTION_ACCESS_DISABLE     (0x01)
+#define  MPU_INSTRUCTION_ACCESS_ENABLE      ((uint8_t)0x00)
+#define  MPU_INSTRUCTION_ACCESS_DISABLE     ((uint8_t)0x01)
 /**
   * @}
   */
@@ -143,8 +143,8 @@ typedef struct
 /** @defgroup CORTEX_MPU_Access_Shareable CORTEX MPU Instruction Access Shareable
   * @{
   */
-#define  MPU_ACCESS_SHAREABLE        (0x01)
-#define  MPU_ACCESS_NOT_SHAREABLE    (0x00)
+#define  MPU_ACCESS_SHAREABLE        ((uint8_t)0x01)
+#define  MPU_ACCESS_NOT_SHAREABLE    ((uint8_t)0x00)
 /**
   * @}
   */
@@ -152,8 +152,8 @@ typedef struct
 /** @defgroup CORTEX_MPU_Access_Cacheable CORTEX MPU Instruction Access Cacheable
   * @{
   */
-#define  MPU_ACCESS_CACHEABLE         (0x01)
-#define  MPU_ACCESS_NOT_CACHEABLE     (0x00)
+#define  MPU_ACCESS_CACHEABLE         ((uint8_t)0x01)
+#define  MPU_ACCESS_NOT_CACHEABLE     ((uint8_t)0x00)
 /**
   * @}
   */
@@ -161,8 +161,8 @@ typedef struct
 /** @defgroup CORTEX_MPU_Access_Bufferable CORTEX MPU Instruction Access Bufferable
   * @{
   */
-#define  MPU_ACCESS_BUFFERABLE         (0x01)
-#define  MPU_ACCESS_NOT_BUFFERABLE     (0x00)
+#define  MPU_ACCESS_BUFFERABLE         ((uint8_t)0x01)
+#define  MPU_ACCESS_NOT_BUFFERABLE     ((uint8_t)0x00)
 /**
   * @}
   */
@@ -170,9 +170,9 @@ typedef struct
 /** @defgroup CORTEX_MPU_TEX_Levels MPU TEX Levels
   * @{
   */
-#define  MPU_TEX_LEVEL0    (0x00)
-#define  MPU_TEX_LEVEL1    (0x01)
-#define  MPU_TEX_LEVEL2    (0x02)
+#define  MPU_TEX_LEVEL0    ((uint8_t)0x00)
+#define  MPU_TEX_LEVEL1    ((uint8_t)0x01)
+#define  MPU_TEX_LEVEL2    ((uint8_t)0x02)
 /**
   * @}
   */
@@ -180,34 +180,34 @@ typedef struct
 /** @defgroup CORTEX_MPU_Region_Size CORTEX MPU Region Size
   * @{
   */
-#define   MPU_REGION_SIZE_32B      (0x04)
-#define   MPU_REGION_SIZE_64B      (0x05)
-#define   MPU_REGION_SIZE_128B     (0x06)
-#define   MPU_REGION_SIZE_256B     (0x07)
-#define   MPU_REGION_SIZE_512B     (0x08)
-#define   MPU_REGION_SIZE_1KB      (0x09)
-#define   MPU_REGION_SIZE_2KB      (0x0A)
-#define   MPU_REGION_SIZE_4KB      (0x0B)
-#define   MPU_REGION_SIZE_8KB      (0x0C)
-#define   MPU_REGION_SIZE_16KB     (0x0D)
-#define   MPU_REGION_SIZE_32KB     (0x0E)
-#define   MPU_REGION_SIZE_64KB     (0x0F)
-#define   MPU_REGION_SIZE_128KB    (0x10)
-#define   MPU_REGION_SIZE_256KB    (0x11)
-#define   MPU_REGION_SIZE_512KB    (0x12)
-#define   MPU_REGION_SIZE_1MB      (0x13)
-#define   MPU_REGION_SIZE_2MB      (0x14)
-#define   MPU_REGION_SIZE_4MB      (0x15)
-#define   MPU_REGION_SIZE_8MB      (0x16)
-#define   MPU_REGION_SIZE_16MB     (0x17)
-#define   MPU_REGION_SIZE_32MB     (0x18)
-#define   MPU_REGION_SIZE_64MB     (0x19)
-#define   MPU_REGION_SIZE_128MB    (0x1A)
-#define   MPU_REGION_SIZE_256MB    (0x1B)
-#define   MPU_REGION_SIZE_512MB    (0x1C)
-#define   MPU_REGION_SIZE_1GB      (0x1D)
-#define   MPU_REGION_SIZE_2GB      (0x1E)
-#define   MPU_REGION_SIZE_4GB      (0x1F)
+#define   MPU_REGION_SIZE_32B      ((uint8_t)0x04)
+#define   MPU_REGION_SIZE_64B      ((uint8_t)0x05)
+#define   MPU_REGION_SIZE_128B     ((uint8_t)0x06)
+#define   MPU_REGION_SIZE_256B     ((uint8_t)0x07)
+#define   MPU_REGION_SIZE_512B     ((uint8_t)0x08)
+#define   MPU_REGION_SIZE_1KB      ((uint8_t)0x09)
+#define   MPU_REGION_SIZE_2KB      ((uint8_t)0x0A)
+#define   MPU_REGION_SIZE_4KB      ((uint8_t)0x0B)
+#define   MPU_REGION_SIZE_8KB      ((uint8_t)0x0C)
+#define   MPU_REGION_SIZE_16KB     ((uint8_t)0x0D)
+#define   MPU_REGION_SIZE_32KB     ((uint8_t)0x0E)
+#define   MPU_REGION_SIZE_64KB     ((uint8_t)0x0F)
+#define   MPU_REGION_SIZE_128KB    ((uint8_t)0x10)
+#define   MPU_REGION_SIZE_256KB    ((uint8_t)0x11)
+#define   MPU_REGION_SIZE_512KB    ((uint8_t)0x12)
+#define   MPU_REGION_SIZE_1MB      ((uint8_t)0x13)
+#define   MPU_REGION_SIZE_2MB      ((uint8_t)0x14)
+#define   MPU_REGION_SIZE_4MB      ((uint8_t)0x15)
+#define   MPU_REGION_SIZE_8MB      ((uint8_t)0x16)
+#define   MPU_REGION_SIZE_16MB     ((uint8_t)0x17)
+#define   MPU_REGION_SIZE_32MB     ((uint8_t)0x18)
+#define   MPU_REGION_SIZE_64MB     ((uint8_t)0x19)
+#define   MPU_REGION_SIZE_128MB    ((uint8_t)0x1A)
+#define   MPU_REGION_SIZE_256MB    ((uint8_t)0x1B)
+#define   MPU_REGION_SIZE_512MB    ((uint8_t)0x1C)
+#define   MPU_REGION_SIZE_1GB      ((uint8_t)0x1D)
+#define   MPU_REGION_SIZE_2GB      ((uint8_t)0x1E)
+#define   MPU_REGION_SIZE_4GB      ((uint8_t)0x1F)
 /**
   * @}
   */
@@ -215,12 +215,12 @@ typedef struct
 /** @defgroup CORTEX_MPU_Region_Permission_Attributes CORTEX MPU Region Permission Attributes 
   * @{
   */
-#define  MPU_REGION_NO_ACCESS      (0x00)
-#define  MPU_REGION_PRIV_RW        (0x01)
-#define  MPU_REGION_PRIV_RW_URO    (0x02)
-#define  MPU_REGION_FULL_ACCESS    (0x03)
-#define  MPU_REGION_PRIV_RO        (0x05)
-#define  MPU_REGION_PRIV_RO_URO    (0x06)
+#define  MPU_REGION_NO_ACCESS      ((uint8_t)0x00)
+#define  MPU_REGION_PRIV_RW        ((uint8_t)0x01)
+#define  MPU_REGION_PRIV_RW_URO    ((uint8_t)0x02)
+#define  MPU_REGION_FULL_ACCESS    ((uint8_t)0x03)
+#define  MPU_REGION_PRIV_RO        ((uint8_t)0x05)
+#define  MPU_REGION_PRIV_RO_URO    ((uint8_t)0x06)
 /**
   * @}
   */
@@ -228,14 +228,14 @@ typedef struct
 /** @defgroup CORTEX_MPU_Region_Number CORTEX MPU Region Number
   * @{
   */
-#define  MPU_REGION_NUMBER0    (0x00)
-#define  MPU_REGION_NUMBER1    (0x01)
-#define  MPU_REGION_NUMBER2    (0x02)
-#define  MPU_REGION_NUMBER3    (0x03)
-#define  MPU_REGION_NUMBER4    (0x04)
-#define  MPU_REGION_NUMBER5    (0x05)
-#define  MPU_REGION_NUMBER6    (0x06)
-#define  MPU_REGION_NUMBER7    (0x07)
+#define  MPU_REGION_NUMBER0    ((uint8_t)0x00)
+#define  MPU_REGION_NUMBER1    ((uint8_t)0x01)
+#define  MPU_REGION_NUMBER2    ((uint8_t)0x02)
+#define  MPU_REGION_NUMBER3    ((uint8_t)0x03)
+#define  MPU_REGION_NUMBER4    ((uint8_t)0x04)
+#define  MPU_REGION_NUMBER5    ((uint8_t)0x05)
+#define  MPU_REGION_NUMBER6    ((uint8_t)0x06)
+#define  MPU_REGION_NUMBER7    ((uint8_t)0x07)
 /**
   * @}
   */
